requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1
PyPDF2==3.0.1
python-docx==1.1.0
openai==1.3.7
redis==5.0.1
celery==5.3.4
supabase==2.0.2
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1 